Created the codeViewConnection and tcpImpl method
In the method codeViewConnection: Initialised EditText txt2 and TextView output1 with findViewById in order to find these views from the layout resource file.
In the method tcpImpl: Created the thread, which automatically generates the method run. In the method run a stream socket is created which connects it to the specified port number at the specified IP adress.